refactor(backend): change directory structure
