+++ apps/backend/src/entities/workout.rs
@@ -3,7 +3,7 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-use crate::fitness::exercise::logic::{WorkoutInformation, WorkoutSummary};
+use crate::fitness::logic::{WorkoutInformation, WorkoutSummary};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
 #[sea_orm(table_name = "workout")]
