--- apps/backend/src/fitness/exercise/mod.rs
@@ -1,2 +1,0 @@
-pub mod logic;
-pub mod resolver;
