+++ apps/backend/src/fitness/logic.rs
