+++ apps/backend/src/fitness/resolver.rs
